Remove Old Method of defining "Analysis Defaults"
This method is very clunky, as it makes generalized assumptions that should really be made on a per-analysis basis. This means that the recommendation is that each analysis or type of analysis should have its own mechanism for resolving defaults.

This can also allow for the creation of a "preset" system, which will be implemetned as an example in the Monte Carlo Analysis Methods